Removed raw tree data from store
